# Conflicts:
#	dwara-core/src/main/java/org/ishafoundation/dwaraapi/storage/archiveformat/ArchivedFile.java

+++ dwara-core/src/main/java/org/ishafoundation/dwaraapi/storage/archiveformat/ArchivedFile.java
@@ -3,7 +3,7 @@ package org.ishafoundation.dwaraapi.storage.archiveformat;
 public class ArchivedFile {
 	private String filePathName;
 	private String linkName;
-	private int archiveBlock; // running total start Archive block of a file
+	private Long archiveBlock; // running total start Archive block of a file
 	private Integer volumeBlock; // running total start Volume block of a file (entire volume level)
 	
 	public String getFilePathName() {
@@ -18,10 +18,10 @@ public class ArchivedFile {
 	public void setLinkName(String linkName) {
 		this.linkName = linkName;
 	}
-	public int getArchiveBlock() {
+	public Long getArchiveBlock() {
 		return archiveBlock;
 	}
-	public void setArchiveBlock(int archiveBlock) {
+	public void setArchiveBlock(Long archiveBlock) {
 		this.archiveBlock = archiveBlock;
 	}
 	public Integer getVolumeBlock() {
