+++ dwara-core/src/main/java/org/ishafoundation/dwaraapi/storage/archiveformat/tar/response/components/File.java
@@ -5,7 +5,7 @@ public class File {
 	private String filePathName;
 	private String linkName;
 	private Long fileSize;
-	private int archiveBlock;
+	private Long archiveBlock;
 //	private int volumeBlock;
 
 
@@ -27,10 +27,10 @@ public class File {
 	public void setFileSize(Long fileSize) {
 		this.fileSize = fileSize;
 	}
-	public int getArchiveBlock() {
+	public Long getArchiveBlock() {
 		return archiveBlock;
 	}
-	public void setArchiveBlock(int archiveBlock) {
+	public void setArchiveBlock(Long archiveBlock) {
 		this.archiveBlock = archiveBlock;
 	}
 //	public int getVolumeBlock() {
